migration end
Окончание работ с миграциями

+++ database/migrations/2025_12_08_114719_create_users_case_statuses_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('users_case_statuses', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('user_id');
+            $table->integer('user_id');
             $table->tinyInteger('case_statuses_id');
             $table->timestamp('created_at');
         });
